Fixed IL2CPP crash by moving `Race/FirstWithIndex` from `Promise<T>` to `Promise`. Split `PromiseStatic.cs` into 3 files for better management.
