+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/PromiseStaticMerge.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ced7977c5e759e47b25eec38f046290
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
