+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/PromiseStaticRace.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4bb591afbb0331d438fdb860d66cf6f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
