+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/Tests/APIs/RaceTests.cs
@@ -474,7 +474,7 @@ namespace ProtoPromiseTests.APIs
             int resultIndex = -1;
             int result = -1;
 
-            Promise<int>.RaceWithIndex(deferred1.Promise, deferred2.Promise)
+            Promise.RaceWithIndex(deferred1.Promise, deferred2.Promise)
                 .Then(cv =>
                 {
                     resultIndex = cv.Item1;
@@ -503,7 +503,7 @@ namespace ProtoPromiseTests.APIs
             int resultIndex = -1;
             int result = -1;
 
-            Promise<int>.RaceWithIndex(deferred1.Promise, deferred2.Promise, deferred3.Promise)
+            Promise.RaceWithIndex(deferred1.Promise, deferred2.Promise, deferred3.Promise)
                 .Then(cv =>
                 {
                     resultIndex = cv.Item1;
@@ -538,7 +538,7 @@ namespace ProtoPromiseTests.APIs
             int resultIndex = -1;
             int result = -1;
 
-            Promise<int>.RaceWithIndex(deferred1.Promise, deferred2.Promise, deferred3.Promise, deferred4.Promise)
+            Promise.RaceWithIndex(deferred1.Promise, deferred2.Promise, deferred3.Promise, deferred4.Promise)
                 .Then(cv =>
                 {
                     resultIndex = cv.Item1;
@@ -578,7 +578,7 @@ namespace ProtoPromiseTests.APIs
             int resultIndex = -1;
             int result = -1;
 
-            Promise<int>.RaceWithIndex(new Promise<int>[] { deferred1.Promise, deferred2.Promise, deferred3.Promise, deferred4.Promise })
+            Promise.RaceWithIndex(new Promise<int>[] { deferred1.Promise, deferred2.Promise, deferred3.Promise, deferred4.Promise })
                 .Then(cv =>
                 {
                     resultIndex = cv.Item1;
